pyproject.toml
[project]
name = "ditto-sionnart-bridge"
version = "0.1.0"
description = "A bridge for real-time radio propagation analysis using Eclipse Ditto and SionnaRT."
readme = "README.md"
requires-python = ">=3.10, <3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "sionna-rt==1.2.1",
    "tensorflow==2.19.0",
    "tensorflow-io-gcs-filesystem<=0.37.1; sys_platform == 'win32'",
    "matplotlib==3.10.7",
    "paho-mqtt",
    "numpy",
    "loguru",
    "PyYAML",
    "pydantic==2.12.4"
]

[tool.uv]
dev-dependencies = []